Check self/internal link hashes
Parsing HTML is now done by htmldoc.Document. The document is parsed and a
slice built of all nodes of interest as well as a hash list. htmltest then
iterates over the interest list and delegates to check functions.

Re-jigged how single documents get tested. As we now need a document store to
resolve hash links we need to fill a store for single doc testing rather than
relying on file system checking. We now fill a store, the same as if we were
checking a directory, resolve the single doc to be tested and test it only.

Create new link check method 'checkHash', called directly for scheme 'self' and
deferred to by checkLink after the reference passes if a fragment is detected.

Add four new tests to validate self hash checking ±ve and internal hash
checking ±ve.

Move attr helper functions to htmldoc.

Re #11