Implement document type checking (DOCTYPE)
New options: CheckDoctype and EnforceHTML5

When CheckDoctype is true we check that a doctype exists.
When EnforceHTML5 is true we check the doctype matches the format for
HTML5.

The golang htmlpackage does most of the heavy lifing for this one.
For example it ignores doctypes which are not the first node in the
document.