Add possibility to ignore based on existing class
Ingoring by tag attribute is really helpful, but eg. docusaurus
does not offer a good way to manipulate its theme. But it offers
a way to always attach a custom class.

With this approach we will reuse the IgnoreTagAttribute also as
a detection for classname.

I do not think it is worth to specify an additional attribute for
this use-case althought the naming might be off now.

Signed-off-by: Simon Schrottner <simon.schrottner@dynatrace.com>